# web/openapi.json
{
 "components": {
  "schemas": {
...
   "CreateItem": {
    "additionalProperties": false,
    "properties": {
     "direction": {
      "default": "my_task",
      "enum": [
       "my_task",
       "my_commitment",
       "delegated",
       "waiting_for",
       "shared",
       "observed",
       "unresolved"
      ],
      "title": "Direction",
      "type": "string"
     },
     "due_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Due At"
     },
     "owner_person_id": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Owner Person Id"
     },
     "title": {
      "maxLength": 300,
      "minLength": 1,
      "title": "Title",
      "type": "string"
     },
     "type": {
      "default": "task",
      "enum": [
       "task",
       "commitment",
       "request",
       "follow_up",
       "deadline"
      ],
      "title": "Type",
      "type": "string"
     }
    },
    "required": [
     "title"
    ],
    "title": "CreateItem",
    "type": "object"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
   "MergeBody": {
    "additionalProperties": false,
    "properties": {
     "into_id": {
      "format": "uuid",
      "title": "Into Id",
      "type": "string"
     }
    },
    "required": [
     "into_id"
    ],
    "title": "MergeBody",
    "type": "object"
   },
...
   "PatchDecision": {
    "additionalProperties": false,
    "properties": {
     "decided_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Decided At"
     },
     "kind": {
      "anyOf": [
       {
        "enum": [
         "decision",
         "open_question"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Kind"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "project_hint": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Hint"
     },
     "rationale": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rationale"
     },
     "statement": {
      "anyOf": [
       {
        "maxLength": 1000,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Statement"
     }
    },
    "title": "PatchDecision",
    "type": "object"
   },
   "PatchItem": {
    "additionalProperties": false,
    "properties": {
     "base_version": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Base Version"
     },
     "counterparty_person_id": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Counterparty Person Id"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "due_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Due At"
     },
     "due_precision": {
      "anyOf": [
       {
        "enum": [
         "datetime",
         "day",
         "week",
         "fuzzy"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Due Precision"
     },
     "due_text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Due Text"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "owner_person_id": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Owner Person Id"
     },
     "project_hint": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Hint"
     },
     "title": {
      "anyOf": [
       {
        "maxLength": 300,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "type": {
      "anyOf": [
       {
        "enum": [
         "task",
         "commitment",
         "request",
         "follow_up",
         "deadline"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Type"
     }
    },
    "title": "PatchItem",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "Executive Context Assistant API",
  "version": "0.0.1"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/decisions": {
   "get": {
    "operationId": "list_decisions_api_v1_decisions_get",
    "parameters": [
     {
      "in": "query",
      "name": "kind",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "decision",
          "open_question"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Kind"
      }
     },
     {
      "in": "query",
      "name": "since",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date-time",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Since"
      }
     },
     {
      "in": "query",
      "name": "cursor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response List Decisions Api V1 Decisions Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Decisions"
   }
  },
  "/api/v1/decisions/{decision_id}": {
   "get": {
    "operationId": "get_decision_api_v1_decisions__decision_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "decision_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Decision Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Decision"
   },
   "patch": {
    "operationId": "patch_decision_api_v1_decisions__decision_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "decision_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Decision Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "If-Match",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "If-Match"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PatchDecision"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Decision"
   }
  },
  "/api/v1/decisions/{decision_id}/{command}": {
   "post": {
    "operationId": "decision_command_api_v1_decisions__decision_id___command__post",
    "parameters": [
     {
      "in": "path",
      "name": "decision_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Decision Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "command",
      "required": true,
      "schema": {
       "enum": [
        "confirm",
        "reject",
        "resolve"
       ],
       "title": "Command",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Decision Command"
   }
  },
  "/api/v1/evidence/{evidence_id}": {
   "get": {
    "operationId": "get_evidence_api_v1_evidence__evidence_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "evidence_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Evidence Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Evidence Api V1 Evidence  Evidence Id  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Evidence"
   }
  },
...
  "/api/v1/work-items": {
   "get": {
    "operationId": "list_work_items_api_v1_work_items_get",
    "parameters": [
     {
      "in": "query",
      "name": "direction",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "my_task",
          "my_commitment",
          "delegated",
          "waiting_for",
          "shared",
          "observed",
          "unresolved"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Direction"
      }
     },
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "default": "open",
       "enum": [
        "open",
        "closed",
        "all"
       ],
       "title": "Status",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "verification",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "suggested",
          "confirmed",
          "rejected",
          "user_created"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Verification"
      }
     },
     {
      "in": "query",
      "name": "person_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "uuid",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Person Id"
      }
     },
     {
      "in": "query",
      "name": "due_before",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date-time",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Due Before"
      }
     },
     {
      "in": "query",
      "name": "sort",
      "required": false,
      "schema": {
       "default": "due",
       "enum": [
        "due",
        "priority",
        "created"
       ],
       "title": "Sort",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "cursor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response List Work Items Api V1 Work Items Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Work Items"
   },
   "post": {
    "operationId": "create_work_item_api_v1_work_items_post",
    "parameters": [
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateItem"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Work Item"
   }
  },
  "/api/v1/work-items/{item_id}": {
   "delete": {
    "operationId": "delete_work_item_api_v1_work_items__item_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "item_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Item Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Work Item"
   },
   "get": {
    "operationId": "get_work_item_api_v1_work_items__item_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "item_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Item Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Work Item"
   },
   "patch": {
    "operationId": "patch_work_item_api_v1_work_items__item_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "item_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Item Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "If-Match",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "If-Match"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PatchItem"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Work Item"
   }
  },
  "/api/v1/work-items/{item_id}/merge": {
   "post": {
    "operationId": "merge_work_item_api_v1_work_items__item_id__merge_post",
    "parameters": [
     {
      "in": "path",
      "name": "item_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Item Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MergeBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Merge Work Item"
   }
  },
  "/api/v1/work-items/{item_id}/{command}": {
   "post": {
    "operationId": "work_item_command_api_v1_work_items__item_id___command__post",
    "parameters": [
     {
      "in": "path",
      "name": "item_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Item Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "command",
      "required": true,
      "schema": {
       "enum": [
        "confirm",
        "reject",
        "complete",
        "reopen",
        "cancel",
        "start"
       ],
       "title": "Command",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Work Item Command"
   }
  },
...
 }
}